third_party/blink/web_tests/fast/forms/calendar-picker/date-picker-in-shadow-ax.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
  <script src="../../../resources/testharness.js"></script>
  <script src="../../../resources/testharnessreport.js"></script>
  <script src="../../../resources/testdriver.js"></script>
  <script src="../../../resources/testdriver-vendor.js"></script>
  <script src="../resources/common.js"></script>
  <script src="../resources/picker-common.js"></script>
  <script src="../calendar-picker/resources/calendar-picker-common.js"></script>


<div id="shadowHost">
  <template shadowrootmode="open">
    <input type="date" id="date0" value="2000-01-02">
  </template>
</div>

<script>
  promise_test(async () => {
    let date = document.querySelector("#shadowHost").shadowRoot.querySelector('#date0');

    await openPicker(date);
    assert_equals(accessibilityController.focusedElement.role, "AXRole: AXGrid", "a11y focus should have moved inside the calendar grid in the control's popup");
  }, 'Tests that a11y focus is sent inside the date picker popup when the control is in a shadow DOM.');
</script>
